pyproject.toml
[project]
name = "fooddesert-review"
version = "0.1.0"
description = ""
authors = [
    {name = "Hyeongchan Bae",email = "83401838+Rlearnchan@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain (>=0.3.17,<0.4.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "langchain-openai (>=0.3.3,<0.4.0)",
    "langchain-chroma (>=0.2.1,<0.3.0)",
    "langgraph (>=0.2.69,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.0,<7.0.0)",
    "pypdf (>=5.2.0,<6.0.0)",
    "streamlit (>=1.41.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
